Update: Move fili fact adapter test to client
